+++ Ass03Solution_NguyenTuanKhai_SE151228/Controllers/PostCategoriesController.cs
@@ -3,27 +3,41 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Ass03Solution_NguyenTuanKhai_SE151228.Models;
+using Microsoft.AspNetCore.SignalR;
+using Microsoft.AspNetCore.Http;
 
 namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
 {
     public class PostCategoriesController : Controller
     {
         private readonly SignalRAssignmentDB03Context _context;
+        private readonly IHubContext<SignalrServer> _signalRHub;
 
-        public PostCategoriesController(SignalRAssignmentDB03Context context)
+        public PostCategoriesController(SignalRAssignmentDB03Context context, IHubContext<SignalrServer> signalRHub)
         {
             _context = context;
+            _signalRHub = signalRHub;
         }
 
         // GET: PostCategories
         public async Task<IActionResult> Index()
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
             return View(await _context.PostCategories.ToListAsync());
         }
 
+        [HttpGet]
+        public IActionResult GetPostCates()
+        {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+            var res = _context.PostCategories.ToList();
+            return Ok(res);
+        }
         // GET: PostCategories/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             if (id == null)
             {
                 return NotFound();
@@ -42,6 +56,8 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         // GET: PostCategories/Create
         public IActionResult Create()
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             return View();
         }
 
@@ -52,10 +68,13 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("CategoryId,CategoryName,Description")] PostCategory postCategory)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             if (ModelState.IsValid)
             {
                 _context.Add(postCategory);
                 await _context.SaveChangesAsync();
+                await _signalRHub.Clients.All.SendAsync("LoadPostCategories");
                 return RedirectToAction(nameof(Index));
             }
             return View(postCategory);
@@ -64,6 +83,8 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         // GET: PostCategories/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             if (id == null)
             {
                 return NotFound();
@@ -84,6 +105,8 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("CategoryId,CategoryName,Description")] PostCategory postCategory)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             if (id != postCategory.CategoryId)
             {
                 return NotFound();
@@ -95,6 +118,7 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
                 {
                     _context.Update(postCategory);
                     await _context.SaveChangesAsync();
+                    await _signalRHub.Clients.All.SendAsync("LoadPostCategories");
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -115,6 +139,8 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         // GET: PostCategories/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             if (id == null)
             {
                 return NotFound();
@@ -135,9 +161,12 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             var postCategory = await _context.PostCategories.FindAsync(id);
             _context.PostCategories.Remove(postCategory);
             await _context.SaveChangesAsync();
+            await _signalRHub.Clients.All.SendAsync("LoadPostCategories");
             return RedirectToAction(nameof(Index));
         }
