+++ Ass03Solution_NguyenTuanKhai_SE151228/Controllers/PostsController.cs
@@ -6,28 +6,46 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Ass03Solution_NguyenTuanKhai_SE151228.Models;
+using Microsoft.AspNetCore.SignalR;
+using Microsoft.Extensions.Configuration;
+using Microsoft.AspNetCore.Http;
 
 namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
 {
     public class PostsController : Controller
     {
         private readonly SignalRAssignmentDB03Context _context;
+        private readonly IHubContext<SignalrServer> _signalRHub;
+        private readonly IConfiguration Configuration;
 
-        public PostsController(SignalRAssignmentDB03Context context)
+        public PostsController(SignalRAssignmentDB03Context context, IHubContext<SignalrServer> signalRHub, IConfiguration configuration)
         {
             _context = context;
+            _signalRHub = signalRHub;
+            Configuration = configuration;
         }
 
         // GET: Posts
         public async Task<IActionResult> Index()
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             var signalRAssignmentDB03Context = _context.Posts.Include(p => p.Author).Include(p => p.Category);
             return View(await signalRAssignmentDB03Context.ToListAsync());
         }
 
+        [HttpGet]
+        public IActionResult GetPosts() {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
+            var res = _context.Posts.ToList();
+            return Ok(res);
+        }
         // GET: Posts/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             if (id == null)
             {
                 return NotFound();
@@ -48,6 +66,8 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         // GET: Posts/Create
         public IActionResult Create()
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             ViewData["AuthorId"] = new SelectList(_context.AppUsers, "UserId", "UserId");
             ViewData["CategoryId"] = new SelectList(_context.PostCategories, "CategoryId", "CategoryName");
             return View();
@@ -60,10 +80,13 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("PostId,CreatedDate,UpdatedDate,Title,Content,PublishStatus,AuthorId,CategoryId")] Post post)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             if (ModelState.IsValid)
             {
                 _context.Add(post);
                 await _context.SaveChangesAsync();
+                await _signalRHub.Clients.All.SendAsync("LoadPosts");
                 return RedirectToAction(nameof(Index));
             }
             ViewData["AuthorId"] = new SelectList(_context.AppUsers, "UserId", "UserId", post.AuthorId);
@@ -74,6 +97,8 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         // GET: Posts/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             if (id == null)
             {
                 return NotFound();
@@ -96,6 +121,8 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("PostId,CreatedDate,UpdatedDate,Title,Content,PublishStatus,AuthorId,CategoryId")] Post post)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             if (id != post.PostId)
             {
                 return NotFound();
@@ -107,6 +134,7 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
                 {
                     _context.Update(post);
                     await _context.SaveChangesAsync();
+                    await _signalRHub.Clients.All.SendAsync("LoadPosts");
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -129,6 +157,8 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         // GET: Posts/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             if (id == null)
             {
                 return NotFound();
@@ -151,9 +181,12 @@ namespace Ass03Solution_NguyenTuanKhai_SE151228.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (HttpContext.Session.GetInt32("id") == null) { HttpContext.Session.SetString("error", "Please login first to access!"); return Redirect("/"); }
+
             var post = await _context.Posts.FindAsync(id);
             _context.Posts.Remove(post);
             await _context.SaveChangesAsync();
+            await _signalRHub.Clients.All.SendAsync("LoadPosts");
             return RedirectToAction(nameof(Index));
         }
